+++ app/src/main/java/com/forrestguice/suntimeswidget/notes/SuntimesNotes3.java
@@ -25,7 +25,9 @@ import android.util.Log;
 
 import com.forrestguice.suntimeswidget.R;
 import com.forrestguice.suntimeswidget.SuntimesUtils;
+import com.forrestguice.suntimeswidget.calculator.SuntimesCalculator;
 import com.forrestguice.suntimeswidget.calculator.SuntimesRiseSetDataset;
+import com.forrestguice.suntimeswidget.settings.AppSettings;
 import com.forrestguice.suntimeswidget.settings.SolarEvents;
 import com.forrestguice.suntimeswidget.settings.WidgetSettings;
 
@@ -65,9 +67,18 @@ public class SuntimesNotes3 implements SuntimesNotes
         this.context = context;
         this.dataset = dataset;
 
+        boolean hasGoldBlue = dataset.calculatorMode().hasRequestedFeature(SuntimesCalculator.FEATURE_GOLDBLUE);
+        boolean enabledGold = AppSettings.loadGoldHourPref(context);
+        boolean enabledBlue = AppSettings.loadBlueHourPref(context);
+
         notesList = new ArrayList<NoteData>();
         for (SolarEvents event : SolarEvents.values())
         {
+            if ((!hasGoldBlue || !enabledGold) && (event.equals(SolarEvents.EVENING_GOLDEN) || event.equals(SolarEvents.MORNING_GOLDEN)))
+                continue;
+            else if ((!hasGoldBlue || !enabledBlue) && (event.equals(SolarEvents.EVENING_BLUE8) || event.equals(SolarEvents.MORNING_BLUE8) || event.equals(SolarEvents.EVENING_BLUE4) || event.equals(SolarEvents.MORNING_BLUE4)))
+                continue;
+
             NoteData note = createNote(event);
             notesList.add(note);
         }
@@ -148,13 +159,13 @@ public class SuntimesNotes3 implements SuntimesNotes
         if (dataset.isCalculated())
         {
             SolarEvents currentNoteMode = WidgetSettings.loadTimeNoteRisePref(context, 0);
-            int currentNote = currentNoteMode.ordinal();
+            int currentNote = getNoteIndex(currentNoteMode);
 
             int nextNote = 0;
-            if (currentNote < SolarEvents.values().length - 1)
+            if (currentNote < notesList.size() - 1)
                 nextNote = currentNote + 1;
 
-            SolarEvents nextNoteMode = SolarEvents.values()[nextNote];
+            SolarEvents nextNoteMode = notesList.get(nextNote).noteMode;
             WidgetSettings.saveTimeNoteRisePref(context, AppWidgetManager.INVALID_APPWIDGET_ID, nextNoteMode);
 
             //Log.d("showNextNote", "... current = " + currentNote + ", next = " + nextNote + ", mode = " + nextNoteMode.name());
@@ -177,15 +188,15 @@ public class SuntimesNotes3 implements SuntimesNotes
         if (dataset.isCalculated())
         {
             SolarEvents currentNoteMode = WidgetSettings.loadTimeNoteRisePref(context, AppWidgetManager.INVALID_APPWIDGET_ID);
-            int currentNote = currentNoteMode.ordinal();
+            int currentNote = getNoteIndex(currentNoteMode);
 
-            int prevNote = SolarEvents.values().length - 1;
+            int prevNote = notesList.size() - 1;
             if (currentNote > 0)
             {
                 prevNote = currentNote - 1;
             }
 
-            SolarEvents prevNoteMode = SolarEvents.values()[prevNote];
+            SolarEvents prevNoteMode = notesList.get(prevNote).noteMode;
             WidgetSettings.saveTimeNoteRisePref(context, AppWidgetManager.INVALID_APPWIDGET_ID, prevNoteMode);
             updateNote(context, dataset.now(), NoteChangedListener.TRANSITION_PREV);
             return true;
@@ -254,17 +265,35 @@ public class SuntimesNotes3 implements SuntimesNotes
                 untilString = context.getString(R.string.until);
                 noteString = context.getString(R.string.until_nauticalTwilight);
                 break;
+            case MORNING_BLUE8:
+                noteIcon = R.drawable.ic_sunrise_large;
+                noteColor = ContextCompat.getColor(context, R.color.sunIcon_color_rising);
+                untilString = context.getString(R.string.until);
+                noteString = context.getString(R.string.until_bluehour);
+                break;
             case MORNING_CIVIL:
                 noteIcon = R.drawable.ic_sunrise_large;
                 noteColor = ContextCompat.getColor(context, R.color.sunIcon_color_rising);
                 untilString = context.getString(R.string.until);
                 noteString = context.getString(R.string.until_civilTwilight);
                 break;
+            case MORNING_BLUE4:
+                noteIcon = R.drawable.ic_sunrise_large;
+                noteColor = ContextCompat.getColor(context, R.color.sunIcon_color_rising);
+                untilString = context.getString(R.string.until_end);
+                noteString = context.getString(R.string.untilEnd_bluehour);
+                break;
             case SUNRISE:
                 noteIcon = R.drawable.ic_sunrise_large;
                 noteColor = ContextCompat.getColor(context, R.color.sunIcon_color_rising);
                 untilString = context.getString(R.string.until);
                 noteString = context.getString(R.string.until_sunrise);
+                break;
+            case MORNING_GOLDEN:
+                noteIcon = R.drawable.ic_sunrise_large;
+                noteColor = ContextCompat.getColor(context, R.color.sunIcon_color_rising);
+                untilString = context.getString(R.string.until);
+                noteString = context.getString(R.string.untilEnd_goldhour);
                 break;
 
             case NOON:
@@ -274,17 +303,35 @@ public class SuntimesNotes3 implements SuntimesNotes
                 noteString = context.getString(R.string.until_noon);
                 break;
 
+            case EVENING_GOLDEN:
+                noteIcon = R.drawable.ic_sunset_large;
+                noteColor = ContextCompat.getColor(context, R.color.sunIcon_color_setting);
+                untilString = context.getString(R.string.until);
+                noteString = context.getString(R.string.until_goldhour);
+                break;
             case SUNSET:
                 noteIcon = R.drawable.ic_sunset_large;
                 noteColor = ContextCompat.getColor(context, R.color.sunIcon_color_setting);
                 untilString = context.getString(R.string.until);
                 noteString = context.getString(R.string.until_sunset);
                 break;
+            case EVENING_BLUE4:
+                noteIcon = R.drawable.ic_sunset_large;
+                noteColor = ContextCompat.getColor(context, R.color.sunIcon_color_setting);
+                untilString = context.getString(R.string.until);
+                noteString = context.getString(R.string.until_bluehour);
+                break;
             case EVENING_CIVIL:
                 noteIcon = R.drawable.ic_sunset_large;
                 noteColor = ContextCompat.getColor(context, R.color.sunIcon_color_setting);
                 untilString = context.getString(R.string.until_end);
                 noteString = context.getString(R.string.untilEnd_civilTwilight);
+                break;
+            case EVENING_BLUE8:
+                noteIcon = R.drawable.ic_sunset_large;
+                noteColor = ContextCompat.getColor(context, R.color.sunIcon_color_setting);
+                untilString = context.getString(R.string.until_end);
+                noteString = context.getString(R.string.untilEnd_bluehour);
                 break;
             case EVENING_NAUTICAL:
                 noteIcon = R.drawable.ic_sunset_large;
@@ -315,30 +362,23 @@ public class SuntimesNotes3 implements SuntimesNotes
         } else {
             switch (event)
             {
-                case MORNING_ASTRONOMICAL:
-                    prefix = context.getString(R.string.until);
-                    break;
+                case MORNING_ASTRONOMICAL:          // until
                 case MORNING_NAUTICAL:
-                    prefix = context.getString(R.string.until);
-                    break;
+                case MORNING_BLUE8:
+                case EVENING_BLUE4:
                 case MORNING_CIVIL:
-                    prefix = context.getString(R.string.until);
-                    break;
                 case SUNRISE:
-                    prefix = context.getString(R.string.until);
-                    break;
                 case NOON:
-                    prefix = context.getString(R.string.until);
-                    break;
+                case EVENING_GOLDEN:
                 case SUNSET:
                     prefix = context.getString(R.string.until);
                     break;
+
+                case MORNING_GOLDEN:               // until_end
                 case EVENING_CIVIL:
-                    prefix = context.getString(R.string.until_end);
-                    break;
+                case EVENING_BLUE8:
+                case MORNING_BLUE4:
                 case EVENING_NAUTICAL:
-                    prefix = context.getString(R.string.until_end);
-                    break;
                 case EVENING_ASTRONOMICAL:
                 default:
                     prefix = context.getString(R.string.until_end);
@@ -366,27 +406,51 @@ public class SuntimesNotes3 implements SuntimesNotes
                 date = dataset.dataNautical.sunriseCalendarToday();
                 dateOther = dataset.dataNautical.sunriseCalendarOther();
                 break;
+            case MORNING_BLUE8:
+                date = dataset.dataBlue8.sunriseCalendarToday();
+                dateOther = dataset.dataBlue8.sunriseCalendarOther();
+                break;
             case MORNING_CIVIL:
                 date = dataset.dataCivil.sunriseCalendarToday();
                 dateOther = dataset.dataCivil.sunriseCalendarOther();
+                break;
+            case MORNING_BLUE4:
+                date = dataset.dataBlue4.sunriseCalendarToday();
+                dateOther = dataset.dataBlue4.sunriseCalendarOther();
                 break;
             case SUNRISE:
                 date = dataset.dataActual.sunriseCalendarToday();
                 dateOther = dataset.dataActual.sunriseCalendarOther();
                 break;
 
+            case MORNING_GOLDEN:
+                date = dataset.dataGold.sunriseCalendarToday();
+                dateOther = dataset.dataGold.sunriseCalendarOther();
+                break;
             case NOON:
                 date = dataset.dataNoon.sunriseCalendarToday();
                 dateOther = dataset.dataNoon.sunriseCalendarOther();
+                break;
+            case EVENING_GOLDEN:
+                date = dataset.dataGold.sunsetCalendarToday();
+                dateOther = dataset.dataGold.sunsetCalendarOther();
                 break;
 
             case SUNSET:
                 date = dataset.dataActual.sunsetCalendarToday();
                 dateOther = dataset.dataActual.sunsetCalendarOther();
                 break;
+            case EVENING_BLUE4:
+                date = dataset.dataBlue4.sunsetCalendarToday();
+                dateOther = dataset.dataBlue4.sunsetCalendarOther();
+                break;
             case EVENING_CIVIL:
                 date = dataset.dataCivil.sunsetCalendarToday();
                 dateOther = dataset.dataCivil.sunsetCalendarOther();
+                break;
+            case EVENING_BLUE8:
+                date = dataset.dataBlue8.sunsetCalendarToday();
+                dateOther = dataset.dataBlue8.sunsetCalendarOther();
                 break;
             case EVENING_NAUTICAL:
                 date = dataset.dataNautical.sunsetCalendarToday();
@@ -459,17 +523,38 @@ public class SuntimesNotes3 implements SuntimesNotes
     @Override
     public void updateNote(Context context, Calendar now, int transition)
     {
-        int choice = WidgetSettings.loadTimeNoteRisePref(context, AppWidgetManager.INVALID_APPWIDGET_ID).ordinal();
-        NoteData chosenNote = notesList.get(choice);
+        SolarEvents choice = WidgetSettings.loadTimeNoteRisePref(context, AppWidgetManager.INVALID_APPWIDGET_ID);
+        NoteData chosenNote = getNote(choice);
 
-        NoteData updatedNote = new NoteData(chosenNote);
-        updateNote(updatedNote, now);
-
-        if (currentNote == null || !currentNote.equals(updatedNote))
+        if (chosenNote != null)
         {
-            //Log.d("updateNote", "changing the note to " + updatedNote.toString() + "[" + choice + "]");
-            setNote(updatedNote, NoteChangedListener.TRANSITION_NEXT);
+            NoteData updatedNote = new NoteData(chosenNote);
+            updateNote(updatedNote, now);
+
+            if (currentNote == null || !currentNote.equals(updatedNote)) {
+                //Log.d("updateNote", "changing the note to " + updatedNote.toString() + "[" + choice + "]");
+                setNote(updatedNote, NoteChangedListener.TRANSITION_NEXT);
+            }
         }
+    }
+
+    public NoteData getNote(SolarEvents event)
+    {
+        int i = getNoteIndex(event);
+        if (i >= 0 && i < notesList.size())
+            return notesList.get(i);
+        else return null;
+    }
+
+    public int getNoteIndex(SolarEvents event)
+    {
+        for (int i=0; i< notesList.size(); i++)
+        {
+            NoteData note = notesList.get(i);
+            if (note.noteMode.equals(event))
+                return i;
+        }
+        return -1;
     }
 
     @Override