+++ app/src/main/java/com/forrestguice/suntimeswidget/settings/WidgetSettings.java
@@ -1,5 +1,5 @@
 /**
-    Copyright (C) 2014 Forrest Guice
+    Copyright (C) 2014-2018 Forrest Guice
     This file is part of SuntimesWidget.
 
     SuntimesWidget is free software: you can redistribute it and/or modify
@@ -791,7 +791,10 @@ public class WidgetSettings
         CIVIL("Civil", "Civil Twilight"),
         NAUTICAL("Nautical", "Nautical Twilight"),
         ASTRONOMICAL("Astronomical", "Astronomical Twilight"),
-        NOON("Noon", "Solar Noon");
+        NOON("Noon", "Solar Noon"),
+        GOLD("Golden", "Golden Hour"),
+        BLUE8("Blue", "Blue Hour"),      // 8 deg; morning start, evening end
+        BLUE4("Blue", "Blue Hour");      // 4 deg; morning end, evening start
 
         public static boolean shortDisplayStrings = false;
         private String longDisplayString;
@@ -801,7 +804,6 @@ public class WidgetSettings
         {
             this.shortDisplayString = shortDisplayString;
             this.longDisplayString = longDisplayString;
-
         }
 
         public String toString()
@@ -847,6 +849,15 @@ public class WidgetSettings
 
             NOON.setDisplayStrings( context.getString(R.string.timeMode_noon_short),
                     context.getString(R.string.timeMode_noon) );
+
+            GOLD.setDisplayStrings( context.getString(R.string.timeMode_golden_short),
+                    context.getString(R.string.timeMode_golden) );
+
+            BLUE8.setDisplayStrings( context.getString(R.string.timeMode_blue8_short),
+                    context.getString(R.string.timeMode_blue8) );
+
+            BLUE4.setDisplayStrings( context.getString(R.string.timeMode_blue4_short),
+                    context.getString(R.string.timeMode_blue4) );
         }
     }